pyproject.toml
[project]
name = "agent-template"
version = "0.1.0"
description = "AI Agent 项目工程模版"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "langchain==1.2.15",
    "langchain-openai==1.1.12",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "python-json-logger>=2.0.0",
    "pydantic-settings>=2.13.1",
    "psycopg[binary]>=3.3.3",
    "langgraph-checkpoint-postgres>=3.0.5",
    "pyyaml>=6.0.3",
]
